staticstack: Move to src/utility
